refactor & add BodySearcher

+++ src/main/java/dev/artiumdominus/creepycrawler/controllers/CrawlerController.java
@@ -8,6 +8,7 @@ import com.google.gson.Gson;
 
 import dev.artiumdominus.creepycrawler.util.IdGenerator;
 import dev.artiumdominus.creepycrawler.models.AnalysisModel;
+import dev.artiumdominus.creepycrawler.models.AnalysisModel.Status;
 import dev.artiumdominus.creepycrawler.models.CrawlModel;
 import dev.artiumdominus.creepycrawler.models.CrawlResponseModel;
 import dev.artiumdominus.creepycrawler.models.ErrorResponseModel;
@@ -16,8 +17,13 @@ import dev.artiumdominus.creepycrawler.workers.CrawlerWorker;
 
 public class CrawlerController {
 
+  private static final Gson gson = new Gson();
+
+  public static Route list = (Request req, Response res) -> {
+    return gson.toJson(AnalysisRepository.list());
+  };
+
   public static Route post = (Request req, Response res) -> {
-    var gson = new Gson();
     CrawlModel crawl = gson.fromJson(req.body(), CrawlModel.class);
 
     if (crawl.keyword == null
@@ -26,7 +32,6 @@ public class CrawlerController {
     {
       res.status(400);
       return gson.toJson(new ErrorResponseModel(404, "field 'keyword' is required (from 4 up to 32 chars)"));
-      // return "{\"status\":400,\"message\":\"field 'keyword' is required (from 4 up to 32 chars)\"}";
     }
 
     var idGenerator = new IdGenerator();
@@ -48,8 +53,6 @@ public class CrawlerController {
 
   public static Route get = (Request req, Response res) ->
   {
-    var gson = new Gson();
-
     var id = req.params("id");
     var analysis = AnalysisRepository.get(id);
 
@@ -58,16 +61,24 @@ public class CrawlerController {
     } else {
       res.status(404);
       return gson.toJson(new ErrorResponseModel(404, "crawl not found: " + id));
-      //return "{\"status\":404,\"message\":\"crawl not found: " + id + "\"}";
     }
   };
 
   public static Route delete = (Request req, Response res) -> {
-    var gson = new Gson();
     var id = req.params("id");
+    var analysis = AnalysisRepository.get(id);
 
-    return AnalysisRepository.delete(id)
-      ? "ok"
-      : gson.toJson(new ErrorResponseModel(404, "crawl not found: " + id));
+    if (analysis != null) {
+      if (analysis.status == Status.DONE) {
+        AnalysisRepository.delete(id);
+        return "ok";
+      } else {
+        res.status(400);
+        return gson.toJson(new ErrorResponseModel(400, "crawl " + id + " is still active."));
+      }
+    } else {
+      res.status(404);
+      return gson.toJson(new ErrorResponseModel(404, "crawl not found: " + id));
+    }
   };
 }
